create works_controller_behavior mixin
