+++ app/controllers/hyrax/publications_controller.rb
@@ -1,17 +1,9 @@
 # frozen_string_literal: true
 module Hyrax
   class PublicationsController < ApplicationController
-    # Adds Hyrax behaviors to the controller.
-    include Hyrax::WorksControllerBehavior
-    include Hyrax::BreadcrumbsForWorks
-    include Spot::AdditionalFormatsForController
-
-    # use our local IiifManifestPresenter for rendering manifest metadata
-    include Spot::ConfigurableIiifManifestPresenter
+    include Spot::WorksControllerBehavior
 
     self.curation_concern_type = ::Publication
-
-    # Use this line if you want to use a custom presenter
     self.show_presenter = Hyrax::PublicationPresenter
   end
 end
